owlVit/requirements.txt
Django
torch
requests
pillow
pyDrive
transformers
boto3
pytz